+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Changed representation of units on `VcsTextField.vue`. Units are now rendered in append slot, the value is not mutated anymore.
 - Adds the functionality to create and display Imprint and Data Protection Information as Link or Markdown Text
 - Adds `markdownHelper.js` to translate markdown into Html
+- Change style of `VcsList.vue` and move select all as action to overflow. Add clear selection action. Update `CollectionComponent.vue` and `CollectionComponentList.vue` according to `VcsList.vue` changes.
 
 ### Ui Behaviour ChangesF